refactor(switch-controller): extract into dedicated crate
Move the switch state controller out of `crates/api` into a new
`carbide-switch-controller` crate as part of the ongoing controller
extraction work.
