+++ crates/switch-controller/Cargo.toml
@@ -1,0 +1,46 @@
+#
+# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+[package]
+name = "carbide-switch-controller"
+version = "0.0.0"
+edition.workspace = true
+license.workspace = true
+authors.workspace = true
+
+[dependencies]
+carbide-api-db = { path = "../api-db", default-features = false }
+carbide-api-model = { path = "../api-model", default-features = false }
+carbide-health-metrics = { path = "../health-metrics" }
+# TODO: probably this dependency is not necessary. It is only use for
+# one function that ultimately calls rms client. Weird dep.
+carbide-site-explorer = { path = "../site-explorer" }
+carbide-secrets = { path = "../secrets" }
+carbide-utils = { path = "../utils", default-features = false }
+carbide-uuid = { path = "../uuid", default-features = false }
+config-version = { path = "../config-version", default-features = false }
+state-controller = { path = "../state-controller" }
+
+async-trait = { workspace = true }
+eyre = { workspace = true }
+mac_address = { workspace = true }
+librms = { workspace = true }
+opentelemetry = { workspace = true }
+sqlx = { workspace = true }
+tracing = { workspace = true }
+
+[lints]
+workspace = true
