+++ crates/switch-controller/src/reprovisioning.rs
@@ -20,11 +20,11 @@
 use carbide_uuid::switch::SwitchId;
 use db::switch as db_switch;
 use model::switch::{ReProvisioningState, Switch, SwitchControllerState};
-
-use crate::state_controller::state_handler::{
+use state_controller::state_handler::{
     StateHandlerContext, StateHandlerError, StateHandlerOutcome,
 };
-use crate::state_controller::switch::context::SwitchStateHandlerContextObjects;
+
+use crate::context::SwitchStateHandlerContextObjects;
 
 /// Handles the ReProvisioning state for a switch.
 pub async fn handle_reprovisioning(
